baselines/acflow/acflow/requirements.txt
python==3.7.6
tensorflow-gpu==1.13.1
numpy==1.18.1
scipy==1.4.1
scikit-learn==0.22.2
pgmpy==0.1.10
networkx==2.4
matplotlib==3.1.3
seaborn==0.10.0
uncertainty-calibration==0.0.7